+++ src/push/topology.rs
@@ -5,15 +5,15 @@ pub struct Topology {}
 impl Topology {
     /// Calculates the Euclidean distance between two index vectors.
     /// Returns None if the vector size dont match.
-    pub fn euclidean_distance(i1: &Vec<usize>, i2: &Vec<usize>) -> Option<f32> {
+    pub fn euclidean_distance(i1: &Vec<usize>, i2: &Vec<usize>) -> Option<f64> {
         if i1.len() != i2.len() {
             None
         } else {
             let mut dist = 0.0;
             for i in 0..i1.len() {
-                dist += (i1[i] as f32 - i2[i] as f32).powf(2.0);
+                dist += (i1[i] as f64 - i2[i] as f64).powf(2.0);
             }
-            Some(f32::sqrt(dist))
+            Some(f64::sqrt(dist))
         }
     }
 
@@ -43,12 +43,12 @@ impl Topology {
         ntotal: &usize,
         ndim: &usize,
         index: &usize,
-        radius: &f32,
+        radius: &f64,
     ) -> Option<IntVector> {
         if *radius < 0.0 || *ndim < 1 || *ntotal < 1 || *index > *ntotal {
             return None;
         }
-        let nedge = f32::ceil((*ntotal as f32).powf(1.0 / *ndim as f32)) as usize;
+        let nedge = f64::ceil((*ntotal as f64).powf(1.0 / *ndim as f64)) as usize;
         if let Some(dindex) = Topology::decompose_index(index, &nedge, ndim) {
             let mut neighbors = vec![];
             for i in 0..*ntotal {
@@ -100,7 +100,7 @@ mod tests {
         );
         assert_eq!(
             Topology::euclidean_distance(&vec![0, 0], &vec![1, 1]),
-            Some(f32::sqrt(2.0))
+            Some(f64::sqrt(2.0))
         );
         assert_eq!(
             Topology::euclidean_distance(&vec![1, 2, 4], &vec![1, 2, 4]),
@@ -123,7 +123,7 @@ mod tests {
             IntVector::new(vec![13, 14, 15])
         );
         assert_eq!(
-            Topology::find_neighbors(&27, &3, &13, &f32::sqrt(3.0)).unwrap(),
+            Topology::find_neighbors(&27, &3, &13, &f64::sqrt(3.0)).unwrap(),
             IntVector::new(vec![
                 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22,
                 23, 24, 25, 26
@@ -138,7 +138,7 @@ mod tests {
     #[test]
     fn find_neighbors_with_envelope() {
         assert_eq!(
-            Topology::find_neighbors(&38, &2, &31, &f32::sqrt(2.0)).unwrap(),
+            Topology::find_neighbors(&38, &2, &31, &f64::sqrt(2.0)).unwrap(),
             IntVector::new(vec![23, 24, 25, 30, 31, 32, 37])
         );
     }
